+++ runtime/ops/mod.rs
@@ -8,7 +8,6 @@ pub mod permissions;
 pub mod process;
 pub mod runtime;
 pub mod signal;
-pub mod spawn;
 pub mod tty;
 mod utils;
 pub mod web_worker;
