+++ runtime/worker.rs
@@ -254,7 +254,6 @@ impl MainWorker {
         options.web_worker_pre_execute_module_cb.clone(),
         options.format_js_error_fn.clone(),
       ),
-      ops::spawn::init(),
       ops::fs_events::init(),
       ops::fs::init::<PermissionsContainer>(),
       deno_io::init(options.stdio),
@@ -269,7 +268,7 @@ impl MainWorker {
       deno_node::init::<PermissionsContainer>(options.npm_resolver),
       ops::os::init(exit_code.clone()),
       ops::permissions::init(),
-      ops::process::init(),
+      ops::process::init_ops(),
       ops::signal::init(),
       ops::tty::init(),
       deno_http::init(),
